This change responds to ticket #121, which says this:
ctwm 3.7 crashes on startup on 64-bit sparcs. The backtrace is:

#0 0x000000004136a16c in _XData32 () from /usr/X11R6/lib/libX11.so.6
#1 0x0000000041390a84 in XChangeProperty () from /usr/X11R6/lib/libX11.so.6
#2 0x0000000000141ea8 in InitGnome () at gnome.c:70
#3 0x00000000001105cc in main (argc=<value optimized out>, 

This is the because the variable `curws' that is passed to XChangeProperty()
is an int not a long. The X code deferences the variable as a long and on
a 64-bit sparc this must be aligned on an 8-byte boundary. Because it is
an int, it is aligned on a 4-byte boundary, causing a bus error.