Fix an completely insane mis-sizing of the occupy window.
When WMgr{Vert,Horiz}ButtonIndent aren't set in the config file, the
occupy window's vertical and horizontal spacing are never initialized.
But they're still used later on for scaling the occupy window, which can
lead to it getting sizes like:

height=1010580661  width=-252644963

which has bad side effects.  So initialize them.